# inline some functions