Running benchmark:

Master:

```
Time Token: 0.4288 seconds
Time RawToken: 0.2157 seconds
```

PR:

```
Time Token: 0.3339 seconds
Time RawToken: 0.1191 seconds
```